Fix bug where derived FCOs may have RelID 0x08000000 after xme import
The issue is that derived FCOs are created all at once with DeriveRootObject,
but, when parsing a non-primary derived, the archetype may not have its RelID
set yet due to a pass_exception (e.g. missing reference target).

The fix is to set the RelID of non-primary deriveds when they are being parsed.

git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@2665 b35830b6-564f-0410-ac5f-b387e50cb686